Refactored the Events to the Common Package
