+++ common/src/main/java/software/bluelib/event/ReloadHandler.java
@@ -11,7 +11,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.packs.resources.CloseableResourceManager;
 import org.jetbrains.annotations.NotNull;
 import software.bluelib.BlueLibConstants;
 import software.bluelib.api.entity.variant.IVariantProvider;
@@ -23,22 +22,25 @@ import software.bluelib.loader.cache.ResourceCache;
 public class ReloadHandler {
 
 	@NotNull
-	private static final List<IVariantProvider> providers = new ArrayList<>();
+	protected static final List<IVariantProvider> providers = new ArrayList<>();
 
-	public static void registerProvider(@NotNull IVariantProvider provider) {
-		providers.add(provider);
+	public static void registerProvider(@NotNull IVariantProvider pProvider) {
+		providers.add(pProvider);
 	}
 
 	public static void onServerStart(@NotNull MinecraftServer pServer) {
 		BlueLibConstants.SCHEDULER = new ScheduledThreadPoolExecutor(1);
 		BlueLibConstants.server = pServer;
+
 		if (providers.isEmpty()) return;
+
 		ResourceCache.registerServerReloadListener(pServer, providers);
 		BaseLogger.log(true, BaseLogLevel.INFO, BlueTranslation.log("variants.loaded"));
 	}
 
-	public static void onReload(@NotNull MinecraftServer pServer, @NotNull CloseableResourceManager pCloseableResourceManager, @NotNull Boolean pBoolean) {
+	public static void onReload(@NotNull MinecraftServer pServer) {
 		if (providers.isEmpty()) return;
+
 		ResourceCache.registerServerReloadListener(pServer, providers);
 		BaseLogger.log(true, BaseLogLevel.INFO, BlueTranslation.log("variants.reloaded"));
 	}
