--- neoforge/src/main/java/software/bluelib/event/ReloadHandler.java
@@ -1,51 +1,0 @@
-/*
- * Copyright (C) 2024 BlueLib Contributors
- *
- * This Source Code Form is subject to the terms of the MIT License.
- * If a copy of the MIT License was not distributed with this file,
- * You can obtain one at https://opensource.org/licenses/MIT.
- */
-package software.bluelib.event;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.ScheduledThreadPoolExecutor;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.event.OnDatapackSyncEvent;
-import net.neoforged.neoforge.event.server.ServerStartingEvent;
-import org.jetbrains.annotations.NotNull;
-import software.bluelib.BlueLibConstants;
-import software.bluelib.api.entity.variant.IVariantProvider;
-import software.bluelib.api.utils.logging.BaseLogLevel;
-import software.bluelib.api.utils.logging.BaseLogger;
-import software.bluelib.internal.BlueTranslation;
-import software.bluelib.loader.cache.ResourceCache;
-
-@EventBusSubscriber(modid = BlueLibConstants.MOD_ID)
-public class ReloadHandler {
-
-	@NotNull
-	private static final List<IVariantProvider> providers = new ArrayList<>();
-
-	public static void registerProvider(@NotNull IVariantProvider pProvider) {
-		providers.add(pProvider);
-	}
-
-	@SubscribeEvent
-	public static void onServerStart(@NotNull ServerStartingEvent pEvent) {
-		BlueLibConstants.SCHEDULER = new ScheduledThreadPoolExecutor(1);
-		BlueLibConstants.server = pEvent.getServer();
-
-		if (providers.isEmpty()) return;
-		ResourceCache.registerServerReloadListener(pEvent.getServer(), providers);
-		BaseLogger.log(true, BaseLogLevel.INFO, BlueTranslation.log("variants.loaded"));
-	}
-
-	@SubscribeEvent
-	public static void onDatapackSync(@NotNull OnDatapackSyncEvent pEvent) {
-		if (providers.isEmpty()) return;
-		ResourceCache.registerServerReloadListener(pEvent.getPlayerList().getServer(), providers);
-		BaseLogger.log(true, BaseLogLevel.INFO, BlueTranslation.log("variants.reloaded"));
-	}
-}
